+++ src/main/UsesCases/IWorkManager.java
@@ -3,7 +3,6 @@ package main.UsesCases;
 import main.Entity.Workable;
 
 public interface IWorkManager {
-
     public void extendWork(Workable work, String extend_date);
     public void changeState(Workable work, String new_statue);
 }
